+++ RentVDB/Models/Customer.cs
@@ -16,7 +16,7 @@ namespace RentVDB.Models
 
         public MembershipType MembershipType { get; set; }
         [Display(Name = "Membership Type")]
-        public byte MembershipTypeId { get; set; }
+        public int MembershipTypeId { get; set; }
 
         [Display(Name = "Date of Birth")]
         [Min18YearsIfAMember]
