+++ RentVDB/Models/MembershipType.cs
@@ -8,15 +8,15 @@ namespace RentVDB.Models
 {
     public class MembershipType
     {
-        public byte Id { get; set; }
+        public int Id { get; set; }
         [Required]
         public string Name { get; set; }
-        public short SignUpFee { get; set; }
-        public byte DurationInMonths { get; set; }
-        public byte DiscountRate { get; set; }
+        public int SignUpFee { get; set; }
+        public int DurationInMonths { get; set; }
+        public int DiscountRate { get; set; }
 
-        public static readonly byte Unknown = 0;
-        public static readonly byte PayAsYouGo = 1;
+        public static readonly int Unknown = 0;
+        public static readonly int PayAsYouGo = 1;
         
     }
 }
